# diagnostic: implement cache size check