@MikeMcQuaid 2018-02-11
@DomT4 How big is your cache currently and how long does this take you to run with/without this patch (run it twice to test the cold/warm IO cache)? Bonus points for @ilovezfs if you want to try that too.