@DomT4 2018-02-11
With an effectively empty cache:

Before this PR:
```
brew doctor  10.41s user 4.71s system 93% cpu 16.107 total
brew doctor  9.42s user 3.37s system 98% cpu 12.963 total
```

After this PR:
```
brew doctor  9.13s user 2.98s system 99% cpu 12.171 total
brew doctor  9.09s user 2.97s system 99% cpu 12.120 total
```

With a cache containing 10 copies of `qt-everywhere-src-5.10.0.tar.xz`, weighing in at 3906MB:

Before this PR:
```
brew doctor  9.77s user 3.23s system 99% cpu 13.120 total
brew doctor  9.46s user 3.13s system 99% cpu 12.692 total
```

After this PR:
```
brew doctor  9.31s user 3.17s system 99% cpu 12.600 total
brew doctor  9.24s user 3.01s system 99% cpu 12.317 total
```

The everything-after-the-first-run results are presumably twisted a bit by _"(run it twice to test the cold/warm IO cache)?"_ but doesn't show a huge slow down.