@ilovezfs 2018-02-11
I think you need something like
```
        return unless cache.to_i > 2000
```
Otherwise it doesn't work for me.