@ilovezfs 2018-02-11
```
irb(main):001:0> "17619" > "2000"
=> false
```