@DomT4 2018-02-11
Oh, it's working here:
```
Please note that these warnings are just used to help the Homebrew maintainers
with debugging if you file an issue. If everything you use Homebrew for is
working fine: please don't worry or file an issue; just ignore this. Thanks!

Warning: Your HOMEBREW_CACHE is using more than 2GB of disk space.
You may wish to consider running `brew cleanup`.
```

Perhaps I didn't expand the cache hard enough during testing.