@ilovezfs 2018-02-11
Yeah sounds like your cache never went over the 2GB mark.