@DomT4 2018-02-11
`3.8G	/usr/local/var/homebrewcache` atm.