@DomT4 2018-02-11
`7.6G	/usr/local/var/homebrewcache` works fine.
` 12G	/usr/local/var/homebrewcache` fails.

So it has issues once we pass the 10000 mark apparently. Will look into that.