@ilovezfs 2018-02-11
and/or `HOMEBREW_CACHE.disk_usage`