@DomT4 2018-02-11
Yeah, we can. In initial testing it was slightly slower than simply throwing out to `du`, and I can repro that locally still:
```
brew doctor  9.44s user 3.04s system 99% cpu 12.554 total
brew doctor  9.49s user 3.11s system 99% cpu 12.694 total
```

But it does work:
```diff
diff --git a/Library/Homebrew/diagnostic.rb b/Library/Homebrew/diagnostic.rb
index 70f3e3881..c38dfedb0 100644
--- a/Library/Homebrew/diagnostic.rb
+++ b/Library/Homebrew/diagnostic.rb
@@ -838,8 +838,8 @@ module Homebrew
       def check_for_large_cache
         return unless HOMEBREW_CACHE.exist?
         return if ENV["CI"] # CI can be expected to have a large cache.
-        cache = Utils.popen_read("du -sm #{HOMEBREW_CACHE}").split.first
-        return unless cache > "2000"
+        cache = HOMEBREW_CACHE.abv.split.last.to_i
+        return unless cache > 2
         <<~EOS
           Your HOMEBREW_CACHE is using more than 2GB of disk space.
           You may wish to consider running `brew cleanup`.
```