@ilovezfs 2018-02-11
```
HOMEBREW_CACHE.disk_usage > 2147483648
```
since 2 is ambiguous depending on the units in the human readable format