@ilovezfs 2018-02-11
Then use `disk_usage_readable(cache_size)` to format it :)