@DomT4 2018-02-11
Not sure if that's still valid/correct. Don't see a problem with long numbers personally but there we go 🙈.