@DomT4 2018-02-11
Specifically:
```
Offenses:

Library/Homebrew/diagnostic.rb:842:36: C: Use underscores(_) as decimal mark and separate every 3 digits with them.
```