@DomT4 2018-02-11
Wasn't sure if shortening it made it less readable at a glance/less friendly.