@DomT4 2018-02-11
Although I guess we could just do `return unless cache_size > 2**31 # ~2GB` if we wanted to retain the instantly-explanatory nature of the full number.