@ilovezfs 2018-02-11
2**31 is the most readable IMO