@ilovezfs 2018-02-11
So in terms of time, on APFS with 4.1GB cache

first doctor run without this PR
real	0m6.313s
user	0m2.081s
sys	0m2.657s

second
real	0m2.850s
user	0m1.897s
sys	0m0.871s

third
real	0m2.924s
user	0m1.964s
sys	0m0.892s

After this PR,

fourth
real	0m5.702s
user	0m2.569s
sys	0m2.185s

fifth
real	0m3.988s
user	0m2.478s
sys	0m1.412s

sixth
real	0m3.838s
user	0m2.424s
sys	0m1.346s