@DomT4 2018-02-11
Before:
```
brew doctor  9.83s user 3.25s system 99% cpu 13.197 total
brew doctor  9.31s user 3.04s system 99% cpu 12.411 total
brew doctor  9.30s user 3.02s system 99% cpu 12.393 total
brew doctor  9.30s user 3.05s system 99% cpu 12.409 total
brew doctor  9.72s user 3.18s system 99% cpu 12.999 total
```

After:
```
brew doctor  10.03s user 3.81s system 90% cpu 15.229 total
brew doctor  9.57s user 3.11s system 99% cpu 12.782 total
brew doctor  9.35s user 3.05s system 99% cpu 12.464 total
brew doctor  9.56s user 3.15s system 99% cpu 12.823 total
brew doctor  9.33s user 3.03s system 99% cpu 12.437 total
```

So a bit of a performance hit from using Ruby rather than a pure `du -sm` comparison, potentially. Not a huge one, and `doctor` is already a slow command, so perhaps not too hideous.