@DomT4 2018-02-12
Managed a mistake so follow up in https://github.com/Homebrew/brew/pull/3791. Thanks for the merge, and apologies for my "oops" 😅.