@ilovezfs 2018-02-13
In light of https://github.com/Homebrew/brew/issues/3798 I think we may want to revert this. It will be an endless source of confusion because `brew cleanup` doesn't guarantee getting under the 2G threshold.