@ilovezfs 2018-02-13
I disagree with the idea that there's anything wrong with a larger cache.